+++ src/atoms/Copyright.js
