+++ src/organisms/Loading.js
@@ -1,5 +1,5 @@
 import React from "react";
-import LoadingAtoms from "../atoms/app-loading";
+import LoadingAtoms from "../atoms/Loading";
 
 export default function Loading() {
   return (
